+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This file documents the most important changes for each released version.
 
 ## [v0.1.1]
+- Switched GUI renderer to wgpu
 - Added automated publish
 
 ## [v0.1.0]